Created Ludl z piezo control class. Created a parent class that contains the main functionality for both the Ludl and MCL piezo voltage control classes.
